requirements.txt
Django>=5.0.0,<5.1.0
python-dotenv>=1.0.0
django-crispy-forms>=2.1
crispy-bootstrap5>=0.7
Pillow>=10.0.0
django-filter>=23.5
django-widget-tweaks>=1.5.0
django-debug-toolbar>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
markdown>=3.5.1 